Add support for PDF

+++ ocr.py
@@ -1,42 +1,74 @@
 from flask import Flask, render_template, request, jsonify
 from werkzeug.utils import secure_filename
 import os
-import sys
 from PIL import Image
 import pytesseract
+import pdf2image
+import pathlib
+from langcodes import Language
 
 __author__ = 'Santhosh Thottingal <santhosh.thottingal@gmail.com>'
-__source__ = ''
+__source__ = 'https://github.com/santhoshtr/tesseract-web'
 
 app = Flask(__name__)
 UPLOAD_FOLDER = './static/uploads'
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 app.config['MAX_CONTENT_LENGTH'] = 10 * 1024 * 1024
+app.config['SUPPORTED_FORMATS'] = ["png","jpeg","jpg","bmp", "pnm", "gif","tiff", "webp", "pdf"]
+
+def pdf_to_img(pdf_file):
+    return pdf2image.convert_from_path(pdf_file)
+
+
+def ocr_core(image:Image, language='en'):
+    text = pytesseract.image_to_string(image, lang=Language.get(language).to_alpha3())
+    return text
+
+def pdf_to_text(pdf_file_path:str, language='en')->str:
+    texts = []
+    images = pdf_to_img(pdf_file_path)
+    for pg, img in enumerate(images):
+        texts.append(ocr_core(img, language))
+
+    return "\n".join(texts)
+
+def get_languages()->dict :
+    languages = {}
+    alpha3codes = pytesseract.get_languages()
+    for code in alpha3codes:
+        language = Language.get(code)
+
+        languages[ language.language ] = language.autonym()
+    return languages
 
 @app.route('/', methods = ['GET', 'POST'])
 def index():
-     languages = pytesseract.get_languages()
-     return render_template("index.html", languages=languages)
+     return render_template("index.html", languages=get_languages())
 
 @app.route('/api/languages', methods = ['GET'])
 def listSupportedLanguages():
-    languages = pytesseract.get_languages()
-    return jsonify(languages=languages)
+    return jsonify(languages=get_languages())
 
 @app.route('/api/ocr', methods = ['POST'])
 def ocr():
     f = request.files['file']
-    language = request.form.get("language", default="eng")
+    language = request.form.get("language", default="en")
     # create a secure filename
     filename = secure_filename(f.filename)
+    file_extension = pathlib.Path(filename).suffix.split('.')[1]
+    if  file_extension not in app.config['SUPPORTED_FORMATS']:
+         return jsonify(error="File format not supported")
 
     # save file to /static/uploads
     filepath = os.path.join(app.config['UPLOAD_FOLDER'],filename)
     f.save(filepath)
 
-
-    # perform OCR on the processed image
-    text = pytesseract.image_to_string(Image.open(filepath), lang=language)
+    if (file_extension == "pdf"):
+        # perform OCR on PDF
+        text = pdf_to_text(filepath, language)
+    else:
+        # perform OCR on the processed image
+        text = ocr_core(Image.open(filepath), language)
 
     # remove the processed image
     os.remove(filepath)
